streaming_anns/requirements.txt
ansicolors==1.1.8
docker==7.1.0
h5py==3.10.0
matplotlib==3.3.4
numpy==1.24.2
pyyaml==6.0
psutil==5.9.4
scipy==1.10.1
scikit-learn
jinja2==3.1.2
pandas==2.0.0